backend/pyproject.toml
[project]
name = "papers-viewer-backend"
version = "0.1.0"
description = "Backend service for extracting content from research papers"
authors = [
    {name = "Luke Shepard"}
]
requires-python = ">=3.11"
dependencies = [
    # FastAPI and web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.18",
    # Supabase client
    "supabase>=2.11.0",
    # AI/ML
    "google-genai>=0.4.0",
    # Research APIs
    "semanticscholar>=0.9.0",
    "perplexipy>=1.1.1",
    # PDF processing
    "pymupdf>=1.24.0",
    "pillow>=10.0.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pytest>=8.4.2",
    # Audio processing
    "pydub>=0.25.1",
    # Video/Audio downloading
    "yt-dlp>=2024.11.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]

[tool.setuptools]
packages = ["lib"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
